+++ cmd/channel-manipulation/join.js
@@ -1,4 +1,3 @@
-#!/usr/bin/env node
 /*!
  * The MIT License (MIT)
  *
@@ -22,8 +21,40 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+// @see http://sharp.dimens.io/en/stable/api-channel/#extractchannel
+
 // Strict mode.
 'use strict'
 
-// Run.
-require('../lib')(process.argv.slice(2))
+// Local modules.
+const baseHandler = require('../../lib/handler')
+const queue = require('../../lib/queue')
+
+// Configure.
+const options = {
+  imageDensity: {
+    desc: 'Integral number representing the DPI for vector images',
+    defaultDescription: 72,
+    type: 'number'
+  },
+  images: { // Hidden option.
+    // desc: 'One or more images',
+    normalize: true,
+    type: 'array'
+  }
+}
+
+// Command handler.
+const handler = (args) => {
+  return queue.push([ 'joinChannel', (sharp) => {
+    return sharp.joinChannel(args.images, { imageDensity: args.imageDensity })
+  }])
+}
+
+// Exports.
+module.exports = {
+  command: 'joinChannel <images..>',
+  describe: 'Join one or more channels to the image',
+  builder: (yargs) => yargs.strict().options(options),
+  handler: baseHandler(handler)
+}