+++ cmd/operations/flop.js
@@ -1,4 +1,3 @@
-#!/usr/bin/env node
 /*!
  * The MIT License (MIT)
  *
@@ -22,8 +21,24 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+// @see http://sharp.dimens.io/en/stable/api-operation/#flop
+
 // Strict mode.
 'use strict'
 
-// Run.
-require('../lib')(process.argv.slice(2))
+// Local modules.
+const baseHandler = require('../../lib/handler')
+const queue = require('../../lib/queue')
+
+// Command handler.
+const handler = (args) => {
+  return queue.push([ 'flop', (sharp) => sharp.flop() ])
+}
+
+// Exports.
+module.exports = {
+  command: 'flop',
+  describe: 'Flop the image about the horizontal X axis',
+  builder: (yargs) => yargs.strict(),
+  handler: baseHandler(handler)
+}