+++ cmd/operations/normalise.js
@@ -1,4 +1,3 @@
-#!/usr/bin/env node
 /*!
  * The MIT License (MIT)
  *
@@ -22,8 +21,25 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+// @see http://sharp.dimens.io/en/stable/api-operation/#normalise
+
 // Strict mode.
 'use strict'
 
-// Run.
-require('../lib')(process.argv.slice(2))
+// Local modules.
+const baseHandler = require('../../lib/handler')
+const queue = require('../../lib/queue')
+
+// Command handler.
+const handler = (args) => {
+  return queue.push([ 'normalise', (sharp) => sharp.normalise() ])
+}
+
+// Exports.
+module.exports = {
+  command: 'normalise',
+  aliases: 'normalize',
+  describe: 'Enhance output image contrast by stretching its luminance to cover the full dynamic range',
+  builder: (yargs) => yargs.strict(),
+  handler: baseHandler(handler)
+}